Changed the powerUp system and added some powerUps.

+++ Assets/Scripts/Projectiles/ProjectilesTypeData.cs
@@ -6,6 +6,5 @@ public class ProjectileTypeData : ScriptableObject
     public string typeName;
     public GameObject prefab;      // Prefab for the projectile visuals
     public float speed;            // Speed of the projectile
-    public int damage;           // Damage dealt by this projectile
-   // public float lifetime;         // How long the projectile exists
+    public float damage;           // Damage dealt by this projectile
 }
